format links
