--- src/scraper/AvailabilityWindow.py
@@ -1,8 +1,0 @@
-class AvailabilityWindow:
-    num_available = 0
-    date = ""
-
-    def __init__(self, num_available, date):
-        self.num_available = num_available
-        self.date = date
-
